Split http1 and http2 converters
Inherit the http2 converter from the http1 one, so we don't have to
duplicate the code to parse the http request.
